### 2022-11-03 #4907 @editorialbot
[PRE REVIEW]: EISPAC - The EIS Python Analysis Code
**Submitting author:** @MJWeberg (<a href="http://orcid.org/0000-0002-4433-4841">Micah J. Weberg</a>)
**Repository:** https://github.com/USNavalResearchLaboratory/eispac
**Branch with paper.md** (empty if default branch):
**Version:** v0.93.3
