### 2022-11-03 #4907 @dfm
@MJWeberg — Thanks for your submission! I'm the managing editor for this JOSS track and I'm working on getting a topic editor assigned to this review. I'll update here with more information soon.

[As you'll see above](https://github.com/openjournals/joss-reviews/issues/4907#issuecomment-1302540865) the PDF for the manuscript is failing to compile because of an invalid ORCID error:

```
Problem with ORCID (0000-0000-0000-0000) for Nicholas Crump. Invalid ORCID (Theoj::Error)
```

In the meantime can you fill in the correct ORCID or just remove it altogether if this author doesn't have one? Thanks!
