### 2022-11-03 e6175eb
simplify date property
