### 2022-11-07 #4907 @mbobra
Thank you very much for agreeing to review, @nabobalis! I really appreciate it.

You can find reviewer guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), together with all the [review criteria](
https://joss.readthedocs.io/en/latest/review_criteria.html). This is what will happen:
* I already assigned you to the pre-review thread (above).
* Once I find a second reviewer, I will begin the review.
* The review will open a new issue on Github. On that issue, you will find a copy of the manuscript, a link to the software repository, and a checklist. During the review, you can check to see if the software repository satisfies the items on the checklist. If not, you can go to the submitting author's software repository, open an issue, and then ask them to address that issue.

Again -- thank you so much.
