### 2022-11-07 #4914 @mbobra
:wave: @nabobalis @PaulJWright Thank you so much for agreeing to review! You can find the article in the comment boxes above ⬆️ , the software repository linked in the first comment box on this issue. To generate your checklist, use the following command:

You can find reviewer guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html), together with all the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html). I think you're good to go. Again, JOSS is an open review process and we encourage communication between the reviewers, the submitting author @MJWeberg, and the editor (myself). Please feel free to ask me questions, I'm always around.

Can you please respond here (or give a thumbs up) so I know you're in the right place and found all the materials?
