### 2022-11-08 USNavalResearchLaboratory/eispac#56 @PaulJWright
Automated build of docs on push/release of version
Looking at the docs:

* the last update was 10 months ago: https://readthedocs.org/projects/eispac/builds/,
* but the most recent (and only) release on pypi was `0.93.3` 4 months ago.

What is the strategy here? Please look into automated building and deployment: https://docs.readthedocs.io/en/stable/features.html.

This is part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/4914
