### 2022-11-08 #4914 @PaulJWright
@mbobra @nabobalis, please let me know your opinions on this, but this open MR seems significant (https://github.com/USNavalResearchLaboratory/eispac/pull/50). I would personally prefer this is merged and tested/documented before any approval
