### 2022-11-08 USNavalResearchLaboratory/eispac#56 @PaulJWright
This is related to https://github.com/USNavalResearchLaboratory/eispac/issues/3
