### 2022-11-08 USNavalResearchLaboratory/eispac#57 @PaulJWright
Address confusion in `ccd_offset.py`
Would you please clarify/solve the comments/questions/confusions in `ccd_offset.py` as reproduced below.

https://github.com/USNavalResearchLaboratory/eispac/blob/571e19095dae6d8d4dded6f7e6566934469d8c4f/eispac/instr/ccd_offset.py#L56
https://github.com/USNavalResearchLaboratory/eispac/blob/571e19095dae6d8d4dded6f7e6566934469d8c4f/eispac/instr/ccd_offset.py#L58
https://github.com/USNavalResearchLaboratory/eispac/blob/571e19095dae6d8d4dded6f7e6566934469d8c4f/eispac/instr/ccd_offset.py#L63

RE Joss review: https://github.com/openjournals/joss-reviews/issues/4914
