### 2022-11-08 USNavalResearchLaboratory/eispac#58 @PaulJWright
Address hardcoded version number
The most recent release on [pypi](https://pypi.org/project/eispac/) was `0.93.3` 4 months ago, however the version number is hardcoded in this repo [here](https://github.com/USNavalResearchLaboratory/eispac/blob/main/eispac/__init__.py). This means that if a user installs through pip, or manually from git, currently they will believe they are on the same version.

I would recommend looking at how this is addressed by `sunpy`.

Example
--------

through pip:

```
(eispac-testing) ➜  eispac git:(main) pip install eispac
(eispac-testing) ➜  eispac git:(main) ipython
Python 3.8.8 (default, Apr 13 2021, 12:59:45)
Type 'copyright', 'credits' or 'license' for more information
IPython 7.22.0 -- An enhanced Interactive Python. Type '?' for help.

In [1]: import eispac

In [2]: eispac.__version__
Out[2]: '0.93.3'
```

... or installing by cloning the repo:
```
(eispac-testing-2) ➜  eispac git:(main) pip install -e .
(eispac-testing-2) ➜  eispac git:(main) ipython
Python 3.8.8 (default, Apr 13 2021, 12:59:45)
Type 'copyright', 'credits' or 'license' for more information
IPython 7.22.0 -- An enhanced Interactive Python. Type '?' for help.

In [1]: import eispac

In [2]: eispac.__version__
Out[2]: '0.93.3'
```

RE: https://github.com/openjournals/joss-reviews/issues/4914
