### 2022-11-08 USNavalResearchLaboratory/eispac#59 @PaulJWright
Discrepancy between copyright notice in LICENSE.txt and `docs/conf.py`
Not a huge deal, but good to be consistent. I would advise changing the LICENSE.txt be consistent

* https://github.com/USNavalResearchLaboratory/eispac/blob/571e19095dae6d8d4dded6f7e6566934469d8c4f/LICENSE.txt#L3
* https://github.com/USNavalResearchLaboratory/eispac/blob/571e19095dae6d8d4dded6f7e6566934469d8c4f/docs/conf.py#L21

RE Joss Review: https://github.com/openjournals/joss-reviews/issues/4914
