### 2022-11-08 USNavalResearchLaboratory/eispac#60 @PaulJWright
Add guidelines on contributing, issue reporting, and seeking support
The JOSS Review (https://github.com/openjournals/joss-reviews/issues/4914) requires a section in the documentation regarding community guidelines.

"Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"

I would recommend adding a section to your sphinx docs, and a reference to this location in your main `README.md`
