### 2022-11-08 USNavalResearchLaboratory/eispac#61 @PaulJWright
Add docstring for `util/scale_intensity.py`
The following function needs a docstring (see the [sphinx docs](https://eispac.readthedocs.io/en/latest/api/eispac.util.scale_intensity.html#eispac.util.scale_intensity)).

https://github.com/USNavalResearchLaboratory/eispac/blob/571e19095dae6d8d4dded6f7e6566934469d8c4f/eispac/util/scale_intensity.py#L6

https://github.com/openjournals/joss-reviews/issues/4914
