### 2022-11-08 USNavalResearchLaboratory/eispac#62 @PaulJWright
Update docstring to be rendered correctly in sphinx
The docstring for `calc_intensity` is not rendered correctly in [sphinx](https://eispac.readthedocs.io/en/latest/api/eispac.util.calc_intensity_range.html)

https://github.com/USNavalResearchLaboratory/eispac/blob/571e19095dae6d8d4dded6f7e6566934469d8c4f/eispac/util/calc_intensity_range.py#L6

https://github.com/openjournals/joss-reviews/issues/4914
