### 2022-11-08 USNavalResearchLaboratory/eispac#63 @PaulJWright
Ensure each class/method/function in the API docs are documented
Please go through the complete Sphinx API docs and ensure that each Class/method/function that is presented has appropriate documentation

e.g.:

* https://eispac.readthedocs.io/en/latest/api/eispac.download.download_db.html
* https://eispac.readthedocs.io/en/latest/api/eispac.download.search.html
* https://eispac.readthedocs.io/en/latest/api/eispac.templates.search.html#eispac.templates.search
* https://eispac.readthedocs.io/en/latest/api/eispac.templates.eis_find_templates.html#

and #61 #62

https://github.com/openjournals/joss-reviews/issues/4914
