### 2022-11-08 #4914 @mbobra
> @mbobra @nabobalis, please let me know your opinions on this, but this open MR seems significant ([USNavalResearchLaboratory/eispac#50](https://github.com/USNavalResearchLaboratory/eispac/pull/50)). I would personally prefer this is merged and tested/documented before any approval.
>
> Please let me know if this is out of scope.

@PaulJWright This is absolutely in scope 👍
