### 2022-11-09 USNavalResearchLaboratory/eispac#64 @nabobalis
Update Install Documentation Page
The current documentation (https://github.com/USNavalResearchLaboratory/eispac/blob/main/docs/guide/00-intro.rst#installation) says its not on PyPi but the readme does.

I think this would check off the Installation instructions part of the documentation for the JOSS review: https://github.com/openjournals/joss-reviews/issues/4914
