### 2022-11-09 b837380
Update docstrings and improve doc links
