### 2022-11-09 USNavalResearchLaboratory/eispac#56 @MJWeberg
Hello,

We have, unfortunately, had issues getting the docs to automatically build correctly. I believe @wtbarnes and @hpwarren both tried to figure it out a while back and confirmed that all the setting _appear_ to be correct. We will take another look at it, but I am not sure what else there is to try at this point.

Until the very recent addition of the Fido client, there was no significant changes to the code base (aside from the PyPi wheel finally getting uploaded). I will work on updating the user's guide next and hopefully have an update about the build process.
