### 2022-11-09 USNavalResearchLaboratory/eispac#64 @MJWeberg
This was an oversight on my part, sorry! The docs unfortunately don't update correctly (as mentioned in #56), but we are looking into it again.
