### 2022-11-09 USNavalResearchLaboratory/eispac#64 @nabobalis
I saw that issue but looking at the actual RST file in the repository, it hasn't been updated.
So when RTD eventually updates, I don't see that changing, unless I am missing something?
