### 2022-11-10 USNavalResearchLaboratory/eispac#56 @PaulJWright
Okay, this will be great to get fixed.

Just so i'm aware what you've tried. Have you set up the webhook, e.g.:
https://docs.readthedocs.io/en/latest/integrations.html#github
