### 2022-11-10 7b522e1
Update `EISMap`  (#50)
* refactor EISMap to not modify metadata in place

* fixup eismap tests

* pin sunpy to >=3.1

* pin to sunpy>=4.0

* bumpy py version in CI

* fix py version number

* unpin sphinx

* simplify date property
