+++ eispac/core/eismap.py
@@ -7,8 +7,8 @@ import numpy as np
 import astropy.units as u
 from astropy.visualization import ImageNormalize, AsinhStretch, LinearStretch
 import sunpy.map
+from sunpy.map.mapbase import SpatialPair
 from sunpy.time import parse_time
-
 
 __all__ = ['EISMap']
 
@@ -23,7 +23,7 @@ class EISMap(sunpy.map.GenericMap):
     contributions from ESA and Norway. Hinode was launched on September 22, 2006
     at 21:36 UTC from the Uchinoura Space Center in Japan and continues to
     operate. EIS observes two wavelength ranges in the extreme ultraviolet,
-    171—212Å and 245—291Å with a spectral resolution of about 22mÅ and a plate
+    171—212 Å and 245—291 Å with a spectral resolution of about 22 mÅ and a plate
     scale of 100 per pixel.
 
     This data structure is designed to hold the fit parameters derived from multi-gaussian
@@ -61,29 +61,10 @@ class EISMap(sunpy.map.GenericMap):
 
         super().__init__(data, header, **kwargs)
 
-        # Validate important keywords and add them in, if missing
-        self.meta['ctype1'] = self.meta.get('ctype1', 'HPLN-TAN')
-        self.meta['ctype2'] = self.meta.get('ctype2', 'HPLT-TAN')
-        self.meta['cunit1'] = self.meta.get('cunit1', 'arcsec')
-        self.meta['cunit2'] = self.meta.get('cunit2', 'arcsec')
-        self.meta['waveunit'] = self.meta.get('waveunit', 'angstrom')
-        self.meta['date-beg'] = self.meta.get('date_beg', self.meta.get('date_obs'))
-        self.meta['date-end'] = self.meta.get('date_end')
-        self.meta['date-obs'] = self.meta.get('date_obs')
-        self.meta['date-avg'] = self.meta.get('date_avg')
-        # NOTE: this block can be removed once sunpy>=3.1 is only supported as
-        # the .date_average property will always be constructed in this way if
-        # date_start and date_end are present
-        if self.meta['date-avg'] is None:
-            if self.meta['date-beg'] is not None and self.meta['date-end'] is not None:
-                timesys = self.meta.get('timesys', 'UTC').lower()
-                start = parse_time(self.meta['date-beg'], scale=timesys)
-                end = parse_time(self.meta['date-end'], scale=timesys)
-                self.meta['date-avg'] = (start + (end - start)/2).isot
-
         # Setup plot settings
         self.plot_settings['aspect'] = self.meta['cdelt2'] / self.meta['cdelt1']
-        # self.plot_settings['interpolation'] = 'kaiser' # KEEP FOR REFERENCE
+        # Adjust colormap and normalization depending on whether the map contains
+        # intensity, velocity, or line width data
         if self.meta['measrmnt'].lower().startswith('int'):
             self.plot_settings['cmap'] = 'Blues_r'
             self.plot_settings['norm'] = ImageNormalize(stretch=AsinhStretch())
@@ -95,6 +76,68 @@ class EISMap(sunpy.map.GenericMap):
         elif self.meta['measrmnt'].lower().startswith('wid'):
             self.plot_settings['cmap'] = 'viridis'
 
+    @property
+    def spatial_units(self):
+        units = self.meta.get('cunit1', 'arcsec'), self.meta.get('cunit2', 'arcsec')
+        return SpatialPair(u.Unit(units[0]), u.Unit(units[1]))
+
+    @property
+    def processing_level(self):
+        return self.meta.get('lvl_num', 3)
+
+    @property
+    def waveunit(self):
+        return u.Unit(self.meta.get('waveunit', 'angstrom'))
+
+    @property
[…]
