+++ setup.cfg
@@ -31,7 +31,7 @@ install_requires =
     matplotlib>=3.1
     h5py>=2.9
     astropy>=3.1
-    sunpy[net,map]>=2.1
+    sunpy[net,map]>=4.0
     ndcube>=2.0.0
     parfive>=1.5
     python-dateutil>=2.8
@@ -41,7 +41,7 @@ test =
     pytest>=4.6.3
     pytest-astropy
 docs =
-    sphinx==4.0.2  # why is this pinned so strictly?
+    sphinx>=4.0.2  # why is this pinned so strictly?
     sphinx-automodapi>=0.13
     sphinx-rtd-theme
 
