+++ setup.cfg
@@ -13,9 +13,9 @@ classifiers =
     License :: OSI Approved :: MIT License
     Programming Language :: Python
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
     Operating System :: OS Independent
 
 [options]
@@ -30,7 +30,7 @@ install_requires =
     scipy>=1.4
     matplotlib>=3.1
     h5py>=2.9
-    astropy>=3.1
+    astropy>=4.2.1
     sunpy[net,map]>=4.0
     ndcube>=2.0.0
     parfive>=1.5
