### 2022-11-11 121a1ea
Update Python and pkg version nums (#66)
* Update Python, SunPy, & AstroPy versions

* Update EISPAC version num
