### 2022-11-12 a776a65
Update RTD and Readme files
