### 2022-11-12 edcf721
Added link to mpfit docs
