### 2022-11-12 384cd7c
Removed all remaining __main__ examples
