+++ eispac/core/fit_spectra.py
@@ -441,60 +441,3 @@ def fit_spectra(inten, template, parinfo=None, wave=None, errs=None, min_points=
         nexp.value = 0
 
     return fit_res
-
-
-if __name__ == '__main__':
-    import matplotlib.pyplot as plt
-    import astropy.units as u
-    # from read_cube import read_cube
-    # from read_template import read_template
-
-    # input data and template files
-    # file_data = './data/eis_20120924_105026.head.h5'
-    file_data = './data/eis_20190404_131513.data.h5'
-    file_template = './templates/eis_template_dir/fe_12_195_119.2c.template.h5'
-
-    # read fit template
-    Fe_XII_195_119 = EISFitTemplate.read_template(file_template)
-
-    # read spectra window
-    raster = read_cube(file_data, Fe_XII_195_119.central_wave)
-
-    # fit profile
-    # wave_coords = raster.axis_world_coords('em.wl')
-    # lower_corner = (300*u.arcsec, 50*u.arcsec, wave_coords[0])
-    # upper_corner = (400*u.arcsec, 150*u.arcsec, wave_coords[-1])
-    # sub_raster = raster.crop_by_coords(lower_corner, upper_corner=upper_corner)
-    sub_raster = raster[0:10, 0:10, :]
-
-    fit_res = fit_spectra(sub_raster, Fe_XII_195_119, ncpu=4)
-
-    # Quick plot raster
-    plot_aspect_ratio = raster.meta['pointing']['y_scale']/raster.meta['pointing']['x_scale']
-    raster.sum_spectra().plot(aspect=plot_aspect_ratio)
-
-    # Plot example fit
-    ex_pxl_coords = [5, 5]
-    # ex_pxl_coords = [4, 8]
-    fit_x, fit_y = fit_res.get_fit_profile(coords=ex_pxl_coords, num_wavelengths=100)
-    c0_fit_x, c0_fit_y = fit_res.get_fit_profile(component=0, coords=ex_pxl_coords,
-                                                 num_wavelengths=100)
-    c1_fit_x, c1_fit_y = fit_res.get_fit_profile(component=1, coords=ex_pxl_coords,
-                                                 num_wavelengths=100)
-    c2_fit_x, c2_fit_y = fit_res.get_fit_profile(component=2, coords=ex_pxl_coords,
-                                                 num_wavelengths=100)
-    sub_data = sub_raster.data[ex_pxl_coords[0], ex_pxl_coords[1], :]
-    sub_wave = sub_raster.wavelength[ex_pxl_coords[0], ex_pxl_coords[1], :]
-    sub_err = sub_raster.uncertainty.array[ex_pxl_coords[0], ex_pxl_coords[1], :]
-
-    fig = plt.figure()
-    profile_subplt = fig.add_subplot(111)
-    profile_subplt.errorbar(sub_wave, sub_data, yerr=sub_err,
-                            ls='', marker='o', color='k')
-    profile_subplt.plot(fit_x, fit_y, color='b')
-    profile_subplt.plot(c0_fit_x, c0_fit_y, color='r')
-    profile_subplt.plot(c1_fit_x, c1_fit_y, color='r', ls='--')
-    profile_subplt.plot(c2_fit_x, c2_fit_y, color='g')
-    profile_subplt.set_xlabel(r'Wavelength [$\AA$]')
-    profile_subplt.set_ylabel('Intensity ['+raster.unit.to_string()+']')
-    plt.show()
