### 2023-04-05 7c65daf
Added versioning using setuptools-scm
