### 2023-04-05 faec6c4
Load default version num from metadata
