+++ eispac/util/rot_xy.py
@@ -9,7 +9,6 @@ from sunpy.time import parse_time
 
 __all__ = ['rot_xy']
 
-
 def rot_xy(xcen, ycen, start_time, end_time):
     """
     Compute the rotation of a point on the Sun using
@@ -36,6 +35,8 @@ def rot_xy(xcen, ycen, start_time, end_time):
     >>> new = rot_xy(0, 0, start_time='2021-JAN-01 00:00', end_time='2021-JAN-01 01:00')
     >>> print(new.Tx, new.Ty)
     9.47188arcsec 0.0809565arcsec
+
+    This is a trivial change
     """
     start_time = parse_time(start_time)
     end_time = parse_time(end_time)
