### 2023-04-24 USNavalResearchLaboratory/eispac#58 @MJWeberg
Finally added automatically updating version numbers.
