### 2023-04-24 USNavalResearchLaboratory/eispac#56 @MJWeberg
After completely rebuilding the docs, we finally fixed the issue with automatic builds on push.
