### 2023-04-24 #4914 @MJWeberg
Hello,

So sorry for the very long delay. The last two issues were minor, but proved to be difficult to find the time to sort out the details. All issues related to this JOSS paper have finally been closed. Thank you for your patience!
