### 2023-05-05 #4914 @mbobra
Congratulations @MJWeberg  -- we're almost ready to publish the paper! Could you please cut a new release of EISPAC, deposit it in Zenodo, and put the link to the Zenodo deposit in this thread? Please make sure that the author list and title on the Zenodo deposit matches the author list and title on the JOSS paper. Then I can move forward with accepting the paper.

See: https://guides.github.com/activities/citable-code/
