### 2023-05-16 2fc4f9d
Fixed bug with reading unicode characters from output files
