### 2023-05-16 84e9433
Cleaned up version numbers for publishing to PyPI
