+++ eispac/_version.py
@@ -1,5 +1,5 @@
 # coding: utf-8
 # file generated by setuptools_scm
 # don't change, don't track in version control
-__version__ = version = '0.94.3.dev10+g0f4e77b'
-__version_tuple__ = version_tuple = (0, 94, 3, 'dev10', 'g0f4e77b')
+__version__ = version = '0.94.5.dev0+gdf7395a.d20230516'
+__version_tuple__ = version_tuple = (0, 94, 5, 'dev0', 'gdf7395a.d20230516')
