### 2023-05-17 bead05c
Fixed type when cleaning up version num
