+++ eispac/_version.py
@@ -1,5 +1,0 @@
-# coding: utf-8
-# file generated by setuptools_scm
-# don't change, don't track in version control
-__version__ = version = '0.94.4.dev0'
-__version_tuple__ = version_tuple = (0, 94, 4, 'dev0')
