### 2023-05-20 aeecb50
journal references, typos, undefined symbol
