+++ docs/joss/eispac_paper.bib
@@ -1,7 +1,7 @@
 @ARTICLE{Culhane:2007,
   author = {{Culhane}, J.~L. and {Harra}, L.~K. and {James}, A.~M. and {Al-Janabi}, K. and {Bradley}, L.~J. and {Chaudry}, R.~A. and {Rees}, K. and {Tandy}, J.~A. and {Thomas}, P. and {Whillock}, M.~C.~R. and {Winter}, B. and {Doschek}, G.~A. and {Korendyke}, C.~M. and {Brown}, C.~M. and {Myers}, S. and {Mariska}, J. and {Seely}, J. and {Lang}, J. and {Kent}, B.~J. and {Shaughnessy}, B.~M. and {Young}, P.~R. and {Simnett}, G.~M. and {Castelli}, C.~M. and {Mahmoud}, S. and {Mapson-Menard}, H. and {Probyn}, B.~J. and {Thomas}, R.~J. and {Davila}, J. and {Dere}, K. and {Windt}, D. and {Shea}, J. and {Hagood}, R. and {Moye}, R. and {Hara}, H. and {Watanabe}, T. and {Matsuzaki}, K. and {Kosugi}, T. and {Hansteen}, V. and {Wikstol}, {\O}.},
   title = "{The EUV Imaging Spectrometer for Hinode}",
-  journal = {\solphys},
+  journal = {Solar Physics},
   year = 2007,
   month = jun,
   volume = {243},
@@ -43,7 +43,7 @@
 @ARTICLE{ssw,
   author = {{Freeland}, S.~L. and {Handy}, B.~N.},
   title = "{Data Analysis with the SolarSoft System}",
-  journal = {\solphys},
+  journal = {Solar Physics},
   year = 1998,
   month = oct,
   volume = {182},
