### 2023-05-20 #4914 @dfm
Thanks @nabobalis and @PaulJWright for your reviews! Thanks also to @mbobra for editing!

I'm going through the final editorial steps and I have just opened a small PR with some edits to the manuscript. Please take a look and then here are the remaining items:

1. @MJWeberg — Can you update the metadata for the Zenodo listing (there should be an "edit" button at the top right hand corner of [this page](https://zenodo.org/record/7949516) if you're logged in) so that the title matches the manuscript?
2. @nabobalis — I see your comment about the "State of the Field" checkbox, but the "Data sharing" and "Reproducibility" boxes also seem to be unchecked. Did you have a reason for omitting those? Thanks!
