### 2023-05-20 #4914 @nabobalis
> 2. @nabobalis — I see your comment about the "State of the Field" checkbox, but the "Data sharing" and "Reproducibility" boxes also seem to be unchecked. Did you have a reason for omitting those? Thanks!

Ah sorry! I actually misread the statement. They should be ticked.
