### 2023-05-23 8c85b08
Silence doctest error due to inconsistent Astropy formatting
