### 2023-05-27 #4914 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
